requirements.txt
Pillow
PyYAML
scikit-image
scikit-learn
tqdm
opencv-python
wget
pydot
configobj
matplotlib-venn
pandas
numpy
optuna